#17 tonejs restructering nk (#36)
* renaming synth to musicCtrl for modularity

* renaming synth to musicCtrl for modularity

* renaming synth to musicCtrl for modularity

* adding new modules

* adding new modules
